Switch to gopkg.in to avoid name issues
In particular error checking[1] is broken if any of the build objects uses
github.com/couchbase/cbcore import path. The API exposes aliases to points
for error names, and those pointers are not actual constant literals
and might vary in different builds (if not static build, it would be kind
of ABI breakage every library build).

[1]: https://github.com/brett19/gocbeventer/blob/05f5ad3d9bd06cb8ae4398dd6e49a91073b14a4f/eventer/watcher.go#L69